add benchmark utils
